Continousclock

+++ OOPP/app/src/main/java/com/example/sauronsarmy/oopp/Monster.java
@@ -10,24 +10,24 @@ import android.media.Image;
 
 public class Monster {
 
-    private double health;
-    private double maxhealth;
+    private int health;
+    private int maxhealth;
     private int gold;
     private int imageref;
 
 
-    public Monster(double health,int gold,int imageref){
+    public Monster(int health, int gold, int imageref){
         this.maxhealth = health;
         this.health = health;
         this.gold = gold;
         this.imageref = imageref;
     }
 
-    public double getHealth(){
+    public int getHealth(){
         return this.health;
     }
 
-    public double getMaxhealth(){
+    public int getMaxhealth(){
         return this.maxhealth;
     }
 
@@ -41,7 +41,7 @@ public class Monster {
     }
 
     //damage a monster, return true if dead
-    public boolean damageMonster(double damage){
+    public boolean damageMonster(int damage){
         this.health -= damage;
         boolean temp = this.health <= 0;
         return temp;
